requirements.txt
fastapi
uvicorn
mysql-connector-python
redis
aioredis
locust